Don't ignore packets with bad time but assume now
So far, the network plugin simply ignores incoming packets that don't contain a "time" part, or carry a timestamp that is too far off. While this is certainly a good measure when looking at computers with a full OS implementation, it is troublesome with microcontrollers or other lean/embedded devices as information sources. For devices without RTC, CMOS battery and only little storage it is hard to keep track of time, especially as local time is often subject to jumps due to daylight savings etc. On the other hand, collectd usually runs on a full-scale POSIX system that keeps time anyway, and latencies for UDP packets in a switched LAN are usually not an issue. This change uses only a simple heuristic, in that if no timestamp is supplied or the time is far off, it is assumed that the measurement was taken "now". To be adopted, probably some kind of configuration would be due.